Move CLI arguments of the `farm` command into `farm` module
