+++ crates/subspace-farmer/src/bin/subspace-farmer/commands.rs
@@ -1,8 +1,7 @@
-mod farm;
+pub(crate) mod farm;
 mod info;
 mod scrub;
 mod shared;
 
-pub(crate) use farm::farm;
 pub(crate) use info::info;
 pub(crate) use scrub::scrub;
